+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.6.0
 
  - Fixed a bug where DNS record lookups never were fired wihout a nameserver
+ - Removed global `--package` option, added `package-manager` property to `package` resource
 
 # v0.5.0